Fixed the issue with iPaymentrepo

+++ src/main/java/com/busticket/dto/response/BookingResponse.java
@@ -12,10 +12,10 @@ import java.time.LocalDateTime;
 @Builder
 public class BookingResponse {
 
-    private Long bookingId;
+    private Integer bookingId;
 
     // Trip summary fields (avoid full TripResponse nesting)
-    private Long tripId;
+    private Integer tripId;
     private String fromCity;
     private String toCity;
     private LocalDateTime departureTime;
