+++ src/main/java/com/busticket/service/impl/CustomerServiceImpl.java
@@ -22,7 +22,7 @@ public class CustomerServiceImpl implements ICustomerService {
 
     @Override
     @Transactional(readOnly = true)
-    public CustomerResponse getCustomerWithAddress(Long customerId) {
+    public CustomerResponse getCustomerWithAddress(Integer customerId) {
         Customer customer = customerRepository.findByIdWithAddress(customerId)
                 .orElseThrow(() ->
                         new ResourceNotFoundException("Customer", "id", customerId));
