+++ src/main/java/com/busticket/service/interfaces/ICustomerService.java
@@ -3,5 +3,5 @@ package com.busticket.service.interfaces;
 import com.busticket.dto.response.CustomerResponse;
 
 public interface ICustomerService {
-    CustomerResponse getCustomerWithAddress(Long customerId);
+    CustomerResponse getCustomerWithAddress(Integer customerId);
 }
